+++ lib/types/csn-extensions.ts
@@ -22,7 +22,8 @@ export enum CdsBuiltinType {
   UInt32 = 'cds.UInt32',
   Integer64 = 'cds.Integer64',
   Decimal = 'cds.Decimal',
-  DecimalFloat = 'cds.DecimalFloat',
+  DecimalFloat = 'cds.Double',
+  // eslint-disable-next-line @typescript-eslint/no-duplicate-enum-values
   Double = 'cds.Double',
   Date = 'cds.Date',
   Time = 'cds.Time',